==> 93.Revision/02.PolypeptideHydrolysis/04.Select_Signals.Rmd <==
---
title: "Signal selecting"
author: "Chao Tang"
date: 'Report created: `r Sys.Date()`'
output: 
  html_document: 
    code_folding: "hide"
    toc: true
    toc_depth: 4
    toc_float: 
      collapsed: false
    number_sections: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, 
                      warning = FALSE, 
                      message = FALSE, 
                      fig.align = 'center')
knitr::opts_knit$set(root.dir = "/mnt/raid61/Personal_data/tangchao/AANanopore")
```

```{r}
library(factoextra)
library(data.table)
library(Biostrings)
library(patchwork)
library(ggplot2)
library(ggrepel)
library(IRanges)
library(dplyr)
```

```{r}
meta <- fread("./data/ChenShanchuan/20231225/meta_20231225.txt", sep = "\t")
meta[, sg_files := paste0("./analysis/81.ABFProcessing/RawSignal/RawSignal_", file_name, ".txt")]
```

# base line distribution

```{r}
RawSig <- lapply(meta[, sg_files], fread)
RawSig <- data.table(file_id = rep(gsub(".txt", "", gsub("RawSignal_", "", basename(meta[, sg_files]))), mapply(nrow, RawSig)), do.call(rbind, RawSig))
RawSig <- merge(RawSig, meta[, .(file_id = file_name, Stage)])
RawSig <- RawSig[Blockade > 0]
```

```{r fig.width=18, fig.height=12}
ggplot(RawSig[BaseMean < 130], aes(x = BaseMean)) + 
  geom_histogram(binwidth = .5) + 
  facet_wrap(~ file_id, scales = "free", nrow = 4)
```

```{r fig.width=18, fig.height=12}
ggplot(RawSig[Stage == "blank"], aes(x = Blockade, y = DwellTime)) + 
  geom_point(size = .1) + 
  scale_y_log10() + 
  facet_grid( ~ file_id, scales = "free") + 
  labs(title = "blank") -> p1
ggplot(RawSig[Stage == "Polypeptide"], aes(x = Blockade, y = DwellTime)) + 
  geom_point(size = .1) + 
  scale_y_log10() + 
  facet_grid( ~ file_id, scales = "free") + 
  labs(title = "Polypeptide") -> p2
ggplot(RawSig[Stage == "Polypeptide + Carboxypeptidase"], aes(x = Blockade, y = DwellTime)) + 
  geom_point(size = .1) + 
  scale_y_log10() + 
  facet_grid( ~ file_id, scales = "free") + 
  labs(title = "Polypeptide + Carboxypeptidase") -> p3

p1 + p2 + p3 + plot_layout(ncol = 1)
```


```{r fig.width=18, fig.height=12}
ggplot(RawSig[Stage == "blank" & Blockade < 0.3], aes(x = Blockade, y = DwellTime)) + 
  geom_point(size = .01) + 
  scale_y_log10() + 
  facet_grid( ~ file_id, scales = "free") + 
  labs(title = "blank") -> p1
ggplot(RawSig[Stage == "Polypeptide" & Blockade < 0.3], aes(x = Blockade, y = DwellTime)) + 
  geom_point(size = .01) + 
  scale_y_log10() + 
  facet_grid( ~ file_id, scales = "free") + 
  labs(title = "Polypeptide") -> p2
ggplot(RawSig[Stage == "Polypeptide + Carboxypeptidase" & Blockade < 0.3], aes(x = Blockade, y = DwellTime)) + 
  geom_point(size = .01) + 
  scale_y_log10() + 
  facet_grid( ~ file_id, scales = "free") + 
  labs(title = "Polypeptide + Carboxypeptidase") -> p3

p1 + p2 + p3 + plot_layout(ncol = 1)
```


# Raw signals

```{r}
Sigs1 <- lapply(list.files("./analysis/93.Revision/02.PolypeptideHydrolysis/02.SelectedL0", full.names = TRUE), fread)
Sigs1 <- do.call(rbind, Sigs1)
Sigs1 <- merge(meta[, .(A = file_id, Stage)], Sigs1)
```

```{r fig.width=18, fig.height=8}
ggplot(Sigs1[Blockade < 0.3 & Stage == "Polypeptide + Carboxypeptidase"], aes(x = Blockade, y = DwellTime)) + 
  geom_point(size = .2) + 
  scale_y_log10() + 
  geom_hline(yintercept = 0.35) + 
  geom_vline(xintercept = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("NSQMF", ""))], Blockade]) + 
  scale_x_continuous(breaks = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("NSQMF", ""))], Blockade], labels = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("NSQMF", ""))], AA]) + 
  facet_wrap( ~ A, nrow = 1) -> p1
ggplot(Sigs1[Blockade < 0.3 & Stage == "Polypeptide"], aes(x = Blockade, y = DwellTime)) + 
  geom_point(size = .2) + 
  scale_y_log10() + 
  geom_hline(yintercept = 0.35) + 
  geom_vline(xintercept = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("NSQMF", ""))], Blockade]) + 
  scale_x_continuous(breaks = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("NSQMF", ""))], Blockade], labels = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("NSQMF", ""))], AA]) + 
  facet_wrap( ~ A, nrow = 1) -> p2
ggplot(Sigs1[Blockade < 0.3 & Stage == "blank"], aes(x = Blockade, y = DwellTime)) + 
  geom_point(size = .2) + 
  scale_y_log10() + 
  geom_hline(yintercept = 0.35) + 
  geom_vline(xintercept = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("NSQMF", ""))], Blockade]) + 
  scale_x_continuous(breaks = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("NSQMF", ""))], Blockade], labels = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("NSQMF", ""))], AA]) + 
  facet_wrap( ~ A, nrow = 1) -> p3
p1 + p2 + p3 + plot_layout(ncol = 1)
```

# Select signal from KNN of euclidean distance

```{r}
dist1 <- readRDS("./analysis/93.Revision/02.PolypeptideHydrolysis/03.SignalDistance/2023_12_25_0005.1_spearman_distance_knn.Rds")
```

```{r fig.width=12, fig.height=4}
sig1 <- KnnSelecting(signals = Sigs1[A == "2023_12_25_0005.1"], dists = dist1, k = 5, noise = 1, index = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("NSQMF", ""))], Blockade])
```

```{r}
dist2 <- readRDS("./analysis/93.Revision/02.PolypeptideHydrolysis/03.SignalDistance/20231225_0009.1_spearman_distance_knn.Rds")
```

```{r fig.width=12, fig.height=4}
sig2 <- KnnSelecting(signals = Sigs1[A == "20231225_0009.1"], dists = dist2, k = 10, noise = 1, index = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("NSQMF", ""))], Blockade])
```

```{r}
dist3 <- readRDS("./analysis/93.Revision/02.PolypeptideHydrolysis/03.SignalDistance/2023_12_24_0004.1_euclidean_distance_knn.Rds")
```

```{r fig.width=24, fig.height=6}
sig3 <- KnnSelecting(signals = Sigs1[A == "2023_12_24_0004.1"], dists = dist3, k = 20, noise = 0, index = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("NSQMF", ""))], Blockade])
```

```{r fig.width=12, fig.height=4}
ggplot(Sigs1[ID %in% c(sig1, sig2, sig3)], aes(x = Blockade, y = DwellTime)) + 
  geom_point(size = .2) + 
  scale_y_log10() + 
  geom_hline(yintercept = 0.35) + 
  geom_vline(xintercept = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("NSQMF", ""))], Blockade]) + 
  scale_x_continuous(breaks = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("NSQMF", ""))], Blockade], labels = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("NSQMF", ""))], AA]) + 
  facet_wrap( ~ A, nrow = 1)
```

# Remove outliers

```{r}
B_file <- unique(meta[!is.na(amino_acid), file_name])
B_file <- paste0("./analysis/81.ABFProcessing/FeatureMatrix/FeatureMatrix_", B_file, ".txt")
B_file <- B_file[file.exists(B_file)]
Sig_FeatureMatrixs <- do.call(rbind, lapply(B_file, fread))
Sig_FeatureMatrixs <- Sig_FeatureMatrixs[ID %in% Sigs1$ID]
FeatureMatrix <- data.frame(Sig_FeatureMatrixs[, grepl("^X", colnames(Sig_FeatureMatrixs)) | colnames(Sig_FeatureMatrixs) %in% c("DwellTime", "DeltaMean", "StageSD", "CurrentSD", "SignalCurrentPercent", "SignalCurrentWidth", "Blockade"), with = F], row.names = Sig_FeatureMatrixs[[1]])
```

```{r}
subtab <- subset.data.frame(FeatureMatrix, row.names(FeatureMatrix) %in% sig1)
res.ds <- stats::dist(subtab, method = "euclidean")
res.hc <- stats::hclust(res.ds, method = "single")
```

```{r}
sig1s <- RemoveOutlier(tree = res.hc, n = 30, Keep = 0.95)
ggplot(Sigs1[ID %in% sig1s], aes(x = Blockade, y = DwellTime, colour = ID %in% sig1s)) + 
  geom_point(size = .5) + 
  scale_y_log10() + 
  scale_colour_manual(values = "#E41A1C") + 
  geom_vline(xintercept = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("NSQMF", ""))], Blockade], colour = "red") + 
  theme_classic() + 
  theme(legend.position = "none") + 
  labs(title = length(sig1s))
```



```{r}
subtab <- subset.data.frame(FeatureMatrix, row.names(FeatureMatrix) %in% sig2)
res.ds <- stats::dist(subtab, method = "euclidean")
res.hc <- stats::hclust(res.ds, method = "single")
```

```{r}
sig2s <- RemoveOutlier(tree = res.hc, n = 30, Keep = 0.95)
ggplot(Sigs1[ID %in% sig2s], aes(x = Blockade, y = DwellTime, colour = ID %in% sig2s)) + 
  geom_point(size = .5) + 
  scale_y_log10() + 
  scale_colour_manual(values = "#E41A1C") + 
  geom_vline(xintercept = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("NSQMF", ""))], Blockade], colour = "red") + 
  theme_classic() + 
  theme(legend.position = "none") + 
  labs(title = length(sig1s))
```



```{r}
subtab <- subset.data.frame(FeatureMatrix, row.names(FeatureMatrix) %in% sig3)
res.ds <- stats::dist(subtab, method = "euclidean")
res.hc <- stats::hclust(res.ds, method = "single")
```

```{r}
sig3s <- RemoveOutlier(tree = res.hc, n = 50, Keep = 0.9)
ggplot(Sigs1[ID %in% sig3s], aes(x = Blockade, y = DwellTime, colour = ID %in% sig3s)) + 
  geom_point(size = .5) + 
  scale_y_log10() + 
  scale_colour_manual(values = "#E41A1C") + 
  geom_vline(xintercept = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("NSQMF", ""))], Blockade], colour = "red") + 
  theme_classic() + 
  theme(legend.position = "none") + 
  labs(title = length(sig1s))
```

```{r}
fwrite(Sigs1[ID %in% c(sig1s, sig2s, sig3s)], "./analysis/93.Revision/02.PolypeptideHydrolysis/04.SelectedSignals/2023_12_24_NSQMF_signals.txt", sep = "\t", quote = FALSE)
```

# Amino acid prediction

```{r}
model1 <- readRDS("./analysis/81.ABFProcessing/SelectedSignals/01.StandardAA/06.Modeling/Version2/L1/SubAAs/ModelAEFLN/RFmodel.Rds")
```

```{r}
selected.signals <- Sigs1[ID %in% c(sig1s, sig2s, sig3s)]
```

```{r}
stopifnot(mean(selected.signals$ID %in% row.names(FeatureMatrix)) == 1)
subtab <- Sig_FeatureMatrixs[ID %in% selected.signals$ID]
pred1 <- data.table(ID = subtab$ID, 
                    Pred1 = as.character(predict(model1, subtab)), 
                    Prob1 = apply(predict(model1, subtab, type = "prob"), 1, max), 
                    Delta1 = apply(predict(model1, subtab, type = "prob"), 1, function(x) - diff(head(sort(x, decreasing = T), 2))))
```

```{r}
NSQMF <- merge(selected.signals, pred1, by = "ID")
```

```{r}
fwrite(NSQMF, "./analysis/93.Revision/02.PolypeptideHydrolysis/04.SelectedSignals/2023_12_24_NSQMF_signals.txt", sep = "\t", quote = FALSE)
```


```{r fig.width=12, fig.height=4}
ggplot(NSQMF, aes(x = Blockade, y = DwellTime, colour = Pred1)) + 
  geom_point(size = .2) + 
  scale_y_log10() + 
  scale_x_continuous(breaks = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("NSQMF", ""))], Blockade], labels = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("NSQMF", ""))], AA]) + 
  facet_wrap( ~ A, nrow = 1)
```


```{r fig.width=12, fig.height=4}
ggplot(NSQMF[Delta1 > 0.9], aes(x = Blockade, y = DwellTime, colour = Pred1)) + 
  geom_point(size = .2) + 
  scale_y_log10() + 
  scale_x_continuous(breaks = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("NSQMF", ""))], Blockade], labels = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("NSQMF", ""))], AA]) + 
  facet_wrap( ~ A, nrow = 1)
```

```{r}
NSQMF[Delta1 > 0.9, .N, .(A, Pred1)]
```

# Valid time
