==> 93.Revision/02.PolypeptideHydrolysis/04.Select_Signals_EAFNL_20231228.Rmd <==
---
title: "Signal selecting"
author: "Chao Tang"
date: 'Report created: `r Sys.Date()`'
output: 
  html_document: 
    code_folding: "hide"
    toc: true
    toc_depth: 4
    toc_float: 
      collapsed: false
    number_sections: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, 
                      warning = FALSE, 
                      message = FALSE, 
                      fig.align = 'center')
knitr::opts_knit$set(root.dir = "/mnt/raid61/Personal_data/tangchao/AANanopore")
```

```{r}
library(factoextra)
library(data.table)
library(Biostrings)
library(patchwork)
library(ggplot2)
library(ggrepel)
library(IRanges)
library(dplyr)
```

```{r}
KnnSelecting <- function(signals, dists, dist_method = "euclidean", k = 20, noise = 0, index = NULL) {
  stopifnot(noise < k)
  if(is(dists, "dist")) {
    knn1 <- dbscan::kNN(dists, k = k)
    knn2 <- apply(knn1$id, 2, function(x) x %in% which(attr(dists, "Labels") %in% grep("^Noise", rownames(knn1$id), value = TRUE)))
  } else {
    if(is(dists, "kNN")) {
      if(k > dists$k) stop("K too big")
      knn1 <- dists
      knn2 <- apply(knn1$id[, 1:k], 2, function(x) x %in% which(grepl("^Noise", rownames(knn1$id))))
    } else {
      knn1 <- dbscan::kNN(dists[[dist_method]], k = k)
      knn2 <- apply(knn1$id, 2, function(x) x %in% which(attr(dists[[dist_method]], "Labels") %in% grep("^Noise", rownames(knn1$id), value = TRUE)))
    }
  }
  
  tu <- row.names(knn1$id)[which(rowSums(knn2) <= noise)]
  # tu <- union(tu, knn2[which(rowSums(knn2) <= noise), 1])
  tu <- grep("^Noise", tu, invert = TRUE, value = TRUE)
  ggplot(signals, aes(x = Blockade, y = DwellTime, colour = !ID %in% tu)) + 
    geom_point(size = .1) + 
    scale_x_continuous(limits = c(0.1, 0.3)) + 
    scale_y_log10() + 
    scale_colour_manual(values = c("#E41A1C", "#377EB8")) + 
    theme_classic() + 
    theme(legend.position = "none") -> p1
  ggplot(signals[ID %in% tu], aes(x = Blockade, y = DwellTime, colour = ID %in% tu)) + 
    geom_point(size = .1) + 
    scale_x_continuous(limits = c(0.1, 0.3)) + 
    scale_y_log10() + 
    scale_colour_manual(values = "#E41A1C") + 
    theme_classic() + 
    theme(legend.position = "none") + 
    labs(title = length(tu)) -> p2
  ggplot(signals[!ID %in% tu], aes(x = Blockade, y = DwellTime, colour = ID %in% tu)) + 
    geom_point(size = .1) + 
    scale_x_continuous(limits = c(0.1, 0.3)) + 
    scale_y_log10() + 
    scale_colour_manual(values = "#377EB8") + 
    theme_classic() + 
    theme(legend.position = "none") -> p3
  if(!is.null(index)) {
    p1 <- p1 + geom_vline(xintercept = index, colour = "red")
    p2 <- p2 + geom_vline(xintercept = index, colour = "red")
    p3 <- p3 + geom_vline(xintercept = index, colour = "red")
  }
  print(p1 + p2 + p3)
  return(tu)
}

RemoveOutlier <- function(tree, n = 10, Keep = 0.95) {
  h <- length(tree$height)
  g <- cutree(tree, h = tree$height[h])
  g <- data.table(tip = names(g), cluster = g)
  
  while (g[, .N, cluster][N >= n, sum(N)/length(tree$labels)] > Keep) {
    h <- h - 1
    g <- cutree(tree, h = tree$height[h])
    g <- data.table(tip = names(g), cluster = g)
  }
  cat(paste0("Keep: ", round(g[, .N, cluster][N >= n, sum(N)/length(tree$labels)] * 100, 2), "%"))
  g[cluster %in% g[, .N, cluster][N >= n, cluster], tip]
}
```
```{r}
KnnSelecting <- function(signals, dists, dist_method = "euclidean", k = 20, noise = 0, index = NULL) {
  stopifnot(noise < k)
  if(is(dists, "dist")) {
    knn1 <- dbscan::kNN(dists, k = k)
    knn2 <- apply(knn1$id, 2, function(x) x %in% which(attr(dists, "Labels") %in% grep("^Noise", rownames(knn1$id), value = TRUE)))
  } else {
    if(is(dists, "kNN")) {
      if(k > dists$k) stop("K too big")
      knn1 <- dists
      knn2 <- apply(knn1$id[, 1:k], 2, function(x) x %in% which(grepl("^Noise", rownames(knn1$id))))
    } else {
      knn1 <- dbscan::kNN(dists[[dist_method]], k = k)
      knn2 <- apply(knn1$id, 2, function(x) x %in% which(attr(dists[[dist_method]], "Labels") %in% grep("^Noise", rownames(knn1$id), value = TRUE)))
    }
  }
  
  tu <- row.names(knn1$id)[which(rowSums(knn2) <= noise)]
  # tu <- union(tu, knn2[which(rowSums(knn2) <= noise), 1])
  tu <- grep("^Noise", tu, invert = TRUE, value = TRUE)
  ggplot(signals, aes(x = Blockade, y = DwellTime, colour = !ID %in% tu)) + 
    geom_point(size = .5) + 
    scale_y_log10() + 
    scale_colour_manual(values = c("#E41A1C", "#377EB8")) + 
    theme_classic() + 
    theme(legend.position = "none") -> p1
  ggplot(signals[ID %in% tu], aes(x = Blockade, y = DwellTime, colour = ID %in% tu)) + 
    geom_point(size = .5) + 
    scale_y_log10() + 
    scale_colour_manual(values = "#E41A1C") + 
    theme_classic() + 
    theme(legend.position = "none") + 
    labs(title = length(tu)) -> p2
  ggplot(signals[!ID %in% tu], aes(x = Blockade, y = DwellTime, colour = ID %in% tu)) + 
    geom_point(size = .5) + 
    scale_y_log10() + 
    scale_colour_manual(values = "#377EB8") + 
    theme_classic() + 
    theme(legend.position = "none") -> p3
  if(!is.null(index)) {
    p1 <- p1 + geom_vline(xintercept = index, colour = "red")
    p2 <- p2 + geom_vline(xintercept = index, colour = "red")
    p3 <- p3 + geom_vline(xintercept = index, colour = "red")
  }
  print(p1 + p2 + p3)
  return(tu)
}

RemoveOutlier <- function(tree, n = 10, Keep = 0.95) {
  h <- length(tree$height)
  g <- cutree(tree, h = tree$height[h])
  g <- data.table(tip = names(g), cluster = g)
  
  while (g[, .N, cluster][N >= n, sum(N)/length(tree$labels)] > Keep) {
    h <- h - 1
    g <- cutree(tree, h = tree$height[h])
    g <- data.table(tip = names(g), cluster = g)
  }
  cat(paste0("Keep: ", round(g[, .N, cluster][N >= n, sum(N)/length(tree$labels)] * 100, 2), "%"))
  g[cluster %in% g[, .N, cluster][N >= n, cluster], tip]
}
```

```{r}
meta <- fread("./data/ChenShanchuan/20231228/meta_20231228.txt", sep = "\t")
meta[, sg_files := paste0("./analysis/81.ABFProcessing/RawSignal/RawSignal_", file_name, ".txt")]
meta[amino_acid == "", amino_acid := NA]
```

# base line distribution

```{r}
RawSig <- lapply(meta[, sg_files], fread)
RawSig <- data.table(file_id = rep(gsub(".txt", "", gsub("RawSignal_", "", basename(meta[, sg_files]))), mapply(nrow, RawSig)), do.call(rbind, RawSig))
RawSig <- RawSig[Blockade > 0]
```

```{r fig.width=18, fig.height=12}
ggplot(RawSig[BaseMean < 150], aes(x = BaseMean)) + 
  geom_histogram(binwidth = .5) + 
  facet_wrap(~ file_id, scales = "free", nrow = 4)
```


```{r fig.width=18, fig.height=4}
ggplot(RawSig[Blockade < 0.3], aes(x = Blockade, y = DwellTime)) + 
  geom_point(size = .01) + 
  scale_y_log10() + 
  facet_grid( ~ file_id, scales = "free") + 
  labs(title = "blank")
```


# Raw signals

```{r}
AABlockade <- lapply(list.files("./analysis/61.SignalSelecting/01.StandardAA/02.SelectedSignals", "_V2.Rds", full.names = T), function(x) readRDS(x)$Summary)
AABlockade <- do.call(rbind, AABlockade)
AABlockade <- AABlockade[State == "State1"]
```

```{r}
Sigs1 <- lapply(list.files("./analysis/93.Revision/02.PolypeptideHydrolysis/02.SelectedL0", full.names = TRUE), fread)
Sigs1 <- do.call(rbind, Sigs1)
Sigs1 <- merge(meta[, .(A = file_id)], Sigs1)
```

```{r fig.width=18, fig.height=8}
ggplot(Sigs1[Blockade < 0.3], aes(x = Blockade, y = DwellTime)) + 
  geom_point(size = .2) + 
  scale_y_log10() + 
  geom_hline(yintercept = 0.35) + 
  geom_vline(xintercept = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("LNFAE", ""))], Blockade]) + 
  scale_x_continuous(breaks = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("LNFAE", ""))], Blockade], labels = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("LNFAE", ""))], AA]) + 
  facet_wrap( ~ A, nrow = 2)
```

```{r}
sig0 <- lapply(meta[!is.na(amino_acid), sg_files], fread)
sig0 <- do.call(rbind, sig0)
B_file <- unique(meta[!is.na(amino_acid), file_name])
B_file <- paste0("./analysis/81.ABFProcessing/FeatureMatrix/FeatureMatrix_", B_file, ".txt")
Sig_FeatureMatrix <- do.call(rbind, lapply(B_file, fread))
Sig_FeatureMatrix <- Sig_FeatureMatrix[ID %in% sig0$ID]
FeatureMatrix <- data.frame(Sig_FeatureMatrix[, grepl("^X", colnames(Sig_FeatureMatrix)) | colnames(Sig_FeatureMatrix) %in% c("DwellTime", "DeltaMean", "StageSD", "CurrentSD", "SignalCurrentPercent", "SignalCurrentWidth", "Blockade"), with = F], row.names = Sig_FeatureMatrix[[1]])
```

# 2023_12_28_0003.1

## Select signal from KNN of euclidean distance

```{r}
dist1 <- readRDS("./analysis/93.Revision/02.PolypeptideHydrolysis/03.SignalDistance/2023_12_28_0003.1_spearman_distance_knn.Rds")
```

```{r fig.width=18, fig.height=4}
sig1 <- KnnSelecting(signals = Sigs1[A == "2023_12_28_0003.1"], dists = dist1, k = 20, noise = 2, index = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("EAFNL", ""))], Blockade])
```

## Remove outliers

```{r}
subtab <- subset.data.frame(FeatureMatrix, row.names(FeatureMatrix) %in% sig1)
res.ds <- stats::dist(subtab, method = "euclidean")
res.hc <- stats::hclust(res.ds, method = "single")
```

```{r}
sig1s <- RemoveOutlier(tree = res.hc, n = 20, Keep = 0.9)
ggplot(Sigs1[ID %in% sig1s], aes(x = Blockade, y = DwellTime, colour = ID %in% sig1s)) + 
  geom_point(size = .5) + 
  scale_y_log10() + 
  scale_colour_manual(values = "#E41A1C") + 
  geom_vline(xintercept = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("LNFAE", ""))], Blockade], colour = "red") + 
  theme_classic() + 
  theme(legend.position = "none") + 
  labs(title = length(sig1s))
```

## Amino acid prediction

```{r}
model1 <- readRDS("./analysis/81.ABFProcessing/SelectedSignals/01.StandardAA/06.Modeling/Version2/L1/SubAAs/ModelAEFLN/RFmodel.Rds")
```

```{r}
selected.signals <- Sigs1[ID %in% c(sig1s)]
```

```{r}
stopifnot(mean(selected.signals$ID %in% row.names(FeatureMatrix)) == 1)
subtab <- Sig_FeatureMatrix[ID %in% selected.signals$ID]
pred1 <- data.table(ID = subtab$ID, 
                    Pred1 = as.character(predict(model1, subtab)), 
                    Prob1 = apply(predict(model1, subtab, type = "prob"), 1, max), 
                    Delta1 = apply(predict(model1, subtab, type = "prob"), 1, function(x) - diff(head(sort(x, decreasing = T), 2))))
```

```{r}
LNFAE <- merge(selected.signals, pred1, by = "ID")
```

```{r}
# fwrite(LNFAE, "./analysis/93.Revision/02.PolypeptideHydrolysis/04.SelectedSignals/2023_12_24_LNFAE_signals.txt", sep = "\t", quote = FALSE)
```

```{r fig.width=6, fig.height=4}
ggplot(LNFAE, aes(x = Blockade, y = DwellTime, colour = Pred1)) + 
  geom_point(size = .8) + 
  scale_y_log10() + 
  scale_x_continuous(breaks = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("LNFAE", ""))], Blockade], labels = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("LNFAE", ""))], AA]) + 
  facet_wrap( ~ A, nrow = 1)
```


```{r fig.width=6, fig.height=4}
ggplot(LNFAE[Delta1 > 0.9], aes(x = Blockade, y = DwellTime, colour = Pred1)) + 
  geom_point(size = 1) + 
  scale_y_log10() + 
  scale_x_continuous(breaks = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("LNFAE", ""))], Blockade], labels = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("LNFAE", ""))], AA]) + 
  facet_wrap( ~ A, nrow = 1)
```


```{r}
Pred_N <- LNFAE[Delta1 > 0.99, .N, .(Pred1, A)]
Pred_N[, A := plyr::mapvalues(Pred1, AMINO_ACID_CODE, names(AMINO_ACID_CODE))]
Pred_N[, A := factor(A, levels = c("E", "A", "F", "N", "L"))]

ggplot(Pred_N, aes(x = A, y = N)) + 
  geom_col() + 
  facet_wrap(~ file_id)

ggplot(Pred_N, aes(x = A, y = N, fill = A)) + 
  geom_col() + 
  scale_y_log10()
```

# 20231228_0004.1

## Select signal from KNN of euclidean distance

```{r}
dist1 <- readRDS("./analysis/93.Revision/02.PolypeptideHydrolysis/03.SignalDistance/20231228_0004.1_spearman_distance_knn.Rds")
```

```{r fig.width=18, fig.height=4}
sig1 <- KnnSelecting(signals = Sigs1[A == "20231228_0004.1"], dists = dist1, k = 20, noise = 1, index = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("EAFNL", ""))], Blockade])
```

## Remove outliers

```{r}
subtab <- subset.data.frame(FeatureMatrix, row.names(FeatureMatrix) %in% sig1)
res.ds <- stats::dist(subtab, method = "euclidean")
res.hc <- stats::hclust(res.ds, method = "single")
```

```{r}
sig1s <- RemoveOutlier(tree = res.hc, n = 20, Keep = 0.95)
ggplot(Sigs1[ID %in% sig1s], aes(x = Blockade, y = DwellTime, colour = ID %in% sig1s)) + 
  geom_point(size = .5) + 
  scale_y_log10() + 
  scale_colour_manual(values = "#E41A1C") + 
  geom_vline(xintercept = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("LNFAE", ""))], Blockade], colour = "red") + 
  theme_classic() + 
  theme(legend.position = "none") + 
  labs(title = length(sig1s))
```

## Amino acid prediction

```{r}
model1 <- readRDS("./analysis/81.ABFProcessing/SelectedSignals/01.StandardAA/06.Modeling/Version2/L1/SubAAs/ModelAEFLN/RFmodel.Rds")
```

```{r}
selected.signals <- Sigs1[ID %in% c(sig1s)]
```

```{r}
stopifnot(mean(selected.signals$ID %in% row.names(FeatureMatrix)) == 1)
subtab <- Sig_FeatureMatrix[ID %in% selected.signals$ID]
pred1 <- data.table(ID = subtab$ID, 
                    Pred1 = as.character(predict(model1, subtab)), 
                    Prob1 = apply(predict(model1, subtab, type = "prob"), 1, max), 
                    Delta1 = apply(predict(model1, subtab, type = "prob"), 1, function(x) - diff(head(sort(x, decreasing = T), 2))))
```

```{r}
LNFAE <- merge(selected.signals, pred1, by = "ID")
```

```{r}
# fwrite(LNFAE, "./analysis/93.Revision/02.PolypeptideHydrolysis/04.SelectedSignals/2023_12_24_LNFAE_signals.txt", sep = "\t", quote = FALSE)
```

```{r fig.width=6, fig.height=4}
ggplot(LNFAE, aes(x = Blockade, y = DwellTime, colour = Pred1)) + 
  geom_point(size = .8) + 
  scale_y_log10() + 
  scale_x_continuous(breaks = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("LNFAE", ""))], Blockade], labels = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("LNFAE", ""))], AA]) + 
  facet_wrap( ~ A, nrow = 1)
```

```{r fig.width=6, fig.height=4}
ggplot(LNFAE[Delta1 > 0.9], aes(x = Blockade, y = DwellTime, colour = Pred1)) + 
  geom_point(size = 1) + 
  scale_y_log10() + 
  scale_x_continuous(breaks = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("LNFAE", ""))], Blockade], labels = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("LNFAE", ""))], AA]) + 
  facet_wrap( ~ A, nrow = 1)
```

```{r}
Pred_N <- LNFAE[Delta1 > 0.9, .N, .(Pred1, A)]
Pred_N[, A := plyr::mapvalues(Pred1, AMINO_ACID_CODE, names(AMINO_ACID_CODE))]
Pred_N[, A := factor(A, levels = c("E", "A", "F", "N", "L"))]

ggplot(Pred_N, aes(x = A, y = N)) + 
  geom_col() + 
  facet_wrap(~ file_id)

ggplot(Pred_N, aes(x = A, y = N, fill = A)) + 
  geom_col() + 
  scale_y_log10()
```

# 20231228_0006.1

## Select signal from KNN of euclidean distance

```{r}
dist1 <- readRDS("./analysis/93.Revision/02.PolypeptideHydrolysis/03.SignalDistance/20231228_0006.1_spearman_distance_knn.Rds")
```

```{r fig.width=18, fig.height=4}
sig1 <- KnnSelecting(signals = Sigs1[A == "20231228_0006.1"], dists = dist1, k = 10, noise = 0, index = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("EAFNL", ""))], Blockade])
```

## Remove outliers

```{r}
subtab <- subset.data.frame(FeatureMatrix, row.names(FeatureMatrix) %in% sig1)
res.ds <- stats::dist(subtab, method = "euclidean")
res.hc <- stats::hclust(res.ds, method = "single")
```

```{r}
sig1s <- RemoveOutlier(tree = res.hc, n = 20, Keep = 0.95)
ggplot(Sigs1[ID %in% sig1s], aes(x = Blockade, y = DwellTime, colour = ID %in% sig1s)) + 
  geom_point(size = .5) + 
  scale_y_log10() + 
  scale_colour_manual(values = "#E41A1C") + 
  geom_vline(xintercept = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("LNFAE", ""))], Blockade], colour = "red") + 
  theme_classic() + 
  theme(legend.position = "none") + 
  labs(title = length(sig1s))
```

## Amino acid prediction

```{r}
model1 <- readRDS("./analysis/81.ABFProcessing/SelectedSignals/01.StandardAA/06.Modeling/Version2/L1/SubAAs/ModelAEFLN/RFmodel.Rds")
```

```{r}
selected.signals <- Sigs1[ID %in% c(sig1s)]
```

```{r}
stopifnot(mean(selected.signals$ID %in% row.names(FeatureMatrix)) == 1)
subtab <- Sig_FeatureMatrix[ID %in% selected.signals$ID]
pred1 <- data.table(ID = subtab$ID, 
                    Pred1 = as.character(predict(model1, subtab)), 
                    Prob1 = apply(predict(model1, subtab, type = "prob"), 1, max), 
                    Delta1 = apply(predict(model1, subtab, type = "prob"), 1, function(x) - diff(head(sort(x, decreasing = T), 2))))
```

```{r}
LNFAE <- merge(selected.signals, pred1, by = "ID")
```

```{r}
# fwrite(LNFAE, "./analysis/93.Revision/02.PolypeptideHydrolysis/04.SelectedSignals/2023_12_24_LNFAE_signals.txt", sep = "\t", quote = FALSE)
```

```{r fig.width=6, fig.height=4}
ggplot(LNFAE, aes(x = Blockade, y = DwellTime, colour = Pred1)) + 
  geom_point(size = .8) + 
  scale_y_log10() + 
  scale_x_continuous(breaks = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("LNFAE", ""))], Blockade], labels = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("LNFAE", ""))], AA]) + 
  facet_wrap( ~ A, nrow = 1)
```

```{r fig.width=6, fig.height=4}
ggplot(LNFAE[Delta1 > 0.9], aes(x = Blockade, y = DwellTime, colour = Pred1)) + 
  geom_point(size = 1) + 
  scale_y_log10() + 
  scale_x_continuous(breaks = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("LNFAE", ""))], Blockade], labels = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("LNFAE", ""))], AA]) + 
  facet_wrap( ~ A, nrow = 1)
```

```{r}
Pred_N <- LNFAE[Delta1 > 0.9, .N, .(Pred1, A)]
Pred_N[, A := plyr::mapvalues(Pred1, AMINO_ACID_CODE, names(AMINO_ACID_CODE))]
Pred_N[, A := factor(A, levels = c("E", "A", "F", "N", "L"))]

ggplot(Pred_N, aes(x = A, y = N)) + 
  geom_col() + 
  facet_wrap(~ file_id)

ggplot(Pred_N, aes(x = A, y = N, fill = A)) + 
  geom_col() + 
  scale_y_log10()
```
