==> 93.Revision/02.PolypeptideHydrolysis/21.EAFNL_V1_from_Model.Rmd <==
---
title: "Figure 5 C"
author: "Chao Tang"
date: 'Report created: `r Sys.Date()`'
output: 
  html_document: 
    code_folding: "hide"
    toc: true
    toc_depth: 4
    toc_float: 
      collapsed: false
    number_sections: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, 
                      warning = FALSE, 
                      message = FALSE, 
                      fig.align = 'center')
knitr::opts_knit$set(root.dir = "/mnt/raid61/Personal_data/tangchao/AANanopore")
```

```{r required packages}
library(Biostrings)
library(data.table)
library(patchwork)
library(ggthemes)
library(parallel)
library(ggplot2)
library(cowplot)
library(ggrepel)
library(ggpubr)
```

```{r}
get_density <- function(x, y, n = 1000, ...) {
  dens <- MASS::kde2d(x, y, n = n, ...)
  ix <- findInterval(x, dens$x)
  iy <- findInterval(y, dens$y)
  ii <- cbind(ix, iy)
  return(dens$z[ii] / max(dens$z[ii]))
}
```

```{r}
aa_cols <- c(RColorBrewer::brewer.pal(n = 8, "Accent")[-c(1, 4, 5, 8)], 
             RColorBrewer::brewer.pal(n = 11, "PiYG")[c(4, 9, 11)], 
             ggsci::pal_aaas()(10), 
             ggsci::pal_locuszoom()(7)[-c(3, 6, 7)], 
             RColorBrewer::brewer.pal(n = 3, name = "Set2")[1:2])
aa_cols <- plyr::mapvalues(aa_cols, "#D43F3AFF", RColorBrewer::brewer.pal(n = 8, name = "BrBG")[1])
names(aa_cols) <- c(AMINO_ACID_CODE[1:20], "CbC", "His1", "His2")
names(aa_cols) <- c(AMINO_ACID_CODE[1:20], "CMC", "His1", "His2")
a_cols <- aa_cols
names(a_cols) <- plyr::mapvalues(names(a_cols), AMINO_ACID_CODE, names(AMINO_ACID_CODE))
```

```{r}
AABlockade <- lapply(list.files("./analysis/61.SignalSelecting/01.StandardAA/02.SelectedSignals", "_V2.Rds", full.names = T), function(x) readRDS(x)$Summary)
AABlockade <- do.call(rbind, AABlockade)
AABlockade <- AABlockade[State == "State1"]
```

```{r}
meta1 <- as.data.table(openxlsx::read.xlsx("./analysis/91.RealTimeHydrolysis/副本meta_ZhangMing.xlsx"))[experiment != "EFG Rep1"]
meta2 <- as.data.table(openxlsx::read.xlsx("./analysis/91.RealTimeHydrolysis/副本meta_ZhangMing2.xlsx"))[experiment != "EFG Rep1"]
setnames(meta2, "X9", "seq")
meta <- rbind(meta1, meta2)[experiment != "EFG Rep1"]
meta[, sg_files := NULL]
meta0 <- meta[is.na(amino_acid)]
meta <- meta[!is.na(amino_acid)]
```

```{r}
# meta <- meta[Stage == "Polypeptide + Hydrolase"]
meta[, sg_files := paste0("./analysis/91.RealTimeHydrolysis/01.SelectedL0/", file_id, ".MainL0.txt")]
meta[, file.exists(sg_files)]
sig1 <- do.call(rbind, lapply(meta$sg_files, fread))
sig1 <- sig1[Blockade > 0 & Blockade < 0.3]
```

```{r}
ggplot(sig1, aes(x = Blockade, y = DwellTime)) + 
  geom_point(size = 0.1) + 
  scale_y_log10() + 
  scale_x_continuous(breaks = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("EAFNL", ""))], Blockade], 
                     labels = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("EAFNL", ""))], AA]) + 
  geom_vline(xintercept = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("EAFNL", ""))], Blockade], color = "red")
```

```{r}
FM0 <- lapply(unique(meta[, paste0("./analysis/81.ABFProcessing/FeatureMatrix/FeatureMatrix_", file_name, ".txt")]), fread)
FM0 <- do.call(rbind, FM0)
stopifnot(all(sig1$ID %in% FM0$ID))
FM0 <- FM0[ID %in% sig1$ID]
```

```{r}
model1 <- readRDS("./analysis/91.RealTimeHydrolysis/00.Models/01.EAFNL/Model_EAFNL_Noise.Rds")
```

```{r}
pred1 <- data.table(ID = FM0$ID, 
                    Pred1 = as.character(predict(model1, FM0)), 
                    Prob1 = apply(predict(model1, FM0, type = "prob"), 1, max), 
                    Delta1 = apply(predict(model1, FM0, type = "prob"), 1, function(x) - diff(head(sort(x, decreasing = T), 2))))
```

```{r}
sig2 <- merge(sig1, pred1, by = "ID")
```

```{r}
ggplot(sig2, aes(x = Blockade, y = DwellTime, colour = Pred1)) + 
  geom_point(size = 0.1) + 
  scale_y_log10() + 
  scale_x_continuous(breaks = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("EAFNL", ""))], Blockade], 
                     labels = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("EAFNL", ""))], AA]) + 
  geom_vline(xintercept = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("EAFNL", ""))], Blockade], color = "red")
```

```{r}
ggplot(sig2[Pred1 != "Noise"], aes(x = Blockade, y = DwellTime, colour = Pred1)) + 
  geom_point(size = 0.1) + 
  scale_y_log10() + 
  scale_x_continuous(breaks = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("EAFNL", ""))], Blockade], 
                     labels = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("EAFNL", ""))], AA]) + 
  geom_vline(xintercept = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("EAFNL", ""))], Blockade], color = "red")
```

```{r}
ggplot(sig2[Pred1 != "Noise" & Delta1 > 0.5], aes(x = Blockade, y = DwellTime, colour = Pred1)) + 
  geom_point(size = 0.1) + 
  scale_y_log10() + 
  scale_x_continuous(breaks = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("EAFNL", ""))], Blockade], 
                     labels = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("EAFNL", ""))], AA]) + 
  geom_vline(xintercept = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("EAFNL", ""))], Blockade], color = "red")
```

```{r}
EAFNL <- sig2[Pred1 != "Noise"]
EAFNL[, file_id := as.character(A)]
EAFNL <- merge(EAFNL, meta[, .(file_id, seq)], by = "file_id")
EAFNL <- data.table(EAFNL, D1 = EAFNL[, get_density(x = Blockade, y = log10(DwellTime))])
EAFNL[, Pred1 := plyr::mapvalues(Pred1, AMINO_ACID_CODE, names(AMINO_ACID_CODE))]
EAFNL[, Pred1 := factor(Pred1, levels = c("E", "A", "F", "N", "L"))]
EAFNL <- EAFNL[Prob1 > 0.7]
EAFNL <- split(EAFNL, EAFNL$Pred1)
EAFNL <- lapply(EAFNL, function(x) data.table(x, D2 = x[, get_density(x = Blockade, y = log10(DwellTime), n = 1000)]))
EAFNL <- do.call(rbind, EAFNL)
EAFNL <- EAFNL[D2 > 0.05]
```

```{r}
# EAFNL <- EAFNL[!file_id %in% c("21808001.2", "21827010.1")]
```

```{r fig.width=8, fig.height=5}
ggplot(EAFNL[sample(.N, .N)]) + 
  geom_point(aes(x = Blockade, y = DwellTime, colour = Pred1, alpha = D1), size = 1) + 
  scale_y_log10() + 
  scale_x_continuous(limits = c(0.1, 0.3), n.breaks = 3) + 
  scale_colour_manual(breaks = names(a_cols[unlist(strsplit("EAFNL", ""))]), values = a_cols[unlist(strsplit("EAFNL", ""))], guide = guide_legend(override.aes = list(size = 2))) + 
  guides(alpha = "none") + 
  facet_wrap(~ seq) + 
  labs(y = "Dwell time (ms)") + 
  ggthemes::theme_calc(base_size = 22) + 
  theme(panel.grid = element_blank(), 
        legend.position = "none",
        legend.title = element_blank(), 
        plot.tag = element_text(size = 20, face = "bold"), 
        plot.background = element_blank()) -> p1
p1
```

```{r}
AADensityPlot1 <- function(x, target, legend = F, adjust = 2) {
  x <- x[Pred1 %in% x[, .N, Pred1][N > 3, Pred1]]
  ggplot() +
    geom_density(data = x, aes(x = Blockade, y = after_stat(count / max(count)), fill = Pred1, colour = Pred1), adjust = adjust, alpha = 0.5) + 
    scale_fill_manual(breaks = names(a_cols), values = a_cols) + 
    scale_colour_manual(breaks = names(a_cols), values = a_cols) + 
    theme_bw(base_size = 22) + 
    theme(legend.title = element_blank(), legend.position = ifelse(legend, "right", "none")) + 
    labs(y = "Relative abundance", x = "blockade") -> p
  pd <- ggplot_build(p)
  tab <- as.data.table(pd$data[[1]])[, .SD[which.max(count), ], colour][order(count, decreasing = T), ][colour %in% a_cols[target]]
  tab <- merge(tab, data.table(AA = names(a_cols), colour = a_cols), by = "colour")
  p + geom_text_repel(data = tab, aes(x = x, y = y * 1.5, label = plyr::mapvalues(AA, AMINO_ACID_CODE, names(AMINO_ACID_CODE)), colour = AA), 
                      size = 6, direction = "y", nudge_y = 0, min.segment.length = unit(100, "cm")) + 
    theme(plot.tag = element_text(size = 20, face = "bold"), panel.grid = element_blank())
}

p2 <- AADensityPlot1(x = EAFNL[seq == "EAFNL"], target = unlist(strsplit("EAFNL", ""))) + facet_wrap(~"EAFNL")
p3 <- AADensityPlot1(x = EAFNL[seq == "LNFAE"], target = unlist(strsplit("EAFNL", ""))) + facet_wrap(~"LNFAE")
```

```{r}
p1 <- p1 + theme(strip.background = element_blank(), strip.text = element_blank(), plot.margin = unit(c(0, 1, 1, 1), "line"))
p2 <- p2 + scale_x_continuous(limits = c(0.1, 0.3)) + theme_void(base_size = 22) + theme(legend.position = "none", plot.margin = unit(c(0, 0, 0, 0), "line"), plot.tag = element_text(face = "bold")) + labs(tag = "c")
p3 <- p3 + scale_x_continuous(limits = c(0.1, 0.3)) + theme_void(base_size = 22) + theme(legend.position = "none", plot.margin = unit(c(0, 0, 0, 1), "line"))
```

```{r fig.width=12, fig.height=6}
design <- "
  12
  33
  33
"
p2 + p3 + p1 + 
  plot_layout(design = design)
```

```{r}
EAFNL[, Pred1 := factor(Pred1, levels = c("E", "A", "F", "N", "L"))]
```

```{r fig.width=6, fig.height=6}
ggplot(EAFNL[, .N, .(seq, Pred1)], aes(x = Pred1, y = N, fill = seq)) + 
  geom_col(position = position_dodge()) + 
  geom_text(aes(x = Pred1, y = N * 1.3, label = N), position = position_dodge(width = 0.9), size = 6) + 
  scale_y_log10() + 
  scale_fill_wsj() + 
  facet_wrap(~ seq) + 
  labs(y = "Number of events", x = "Identified amino acid", tag = "d") + 
  theme_base(base_size = 22) + 
  theme(plot.background = element_blank(), 
        # axis.title.x = element_blank(), 
        plot.tag = element_text(face = "bold"), 
        legend.position = "none") -> p4
p4
```


```{r}
IntervalTime <- as.data.table(openxlsx::read.xlsx("./analysis/00.FigureTables/Rebuttal/Nov12/Figures/Data/Fig2C.xlsx"))
setnames(IntervalTime, "AminoAcid", "AA")
IntervalTime[, Class := factor(Class, levels = c("Charged", "Nonpolar", "Polar"))]
od <- IntervalTime[, median(Frequency), .(Class, AA)][order(Class, V1), AA]
IntervalTime[, AA := factor(AA, levels = od)]
IntervalTime <- IntervalTime[, .(Mean = mean(Frequency), Median = median(Frequency)), AA]
IntervalTime[, A := plyr::mapvalues(AA, AMINO_ACID_CODE, names(AMINO_ACID_CODE))]
```

```{r}
Freq_N <- EAFNL[, .N, .(seq, Pred1)]
Freq_N <- merge(Freq_N, IntervalTime, by.x = "Pred1", by.y = "A")
```

```{r fig.width=6, fig.height=6}
ggplot(Freq_N, aes(x = Pred1, y = N/Mean, fill = seq)) + 
  geom_col(position = position_dodge()) + 
  geom_text(aes(x = Pred1, y = N/Mean * 1.3, label = round(N/Mean)), position = position_dodge(width = 0.9), size = 6) + 
  scale_y_log10() + 
  scale_fill_wsj() + 
  facet_wrap(~ seq) + 
  labs(y = "Number of events", x = "Identified amino acid", tag = "d") + 
  theme_base(base_size = 22) + 
  theme(plot.background = element_blank(), 
        # axis.title.x = element_blank(), 
        plot.tag = element_text(face = "bold"), 
        legend.position = "none")
```


```{r fig.width=6, fig.height=6}
ggplot(Freq_N, aes(x = Pred1, y = N/Median, fill = seq)) + 
  geom_col(position = position_dodge()) + 
  geom_text(aes(x = Pred1, y = N/Median * 1.3, label = round(N/Median)), position = position_dodge(width = 0.9), size = 6) + 
  scale_y_log10() + 
  scale_fill_wsj() + 
  facet_wrap(~ seq) + 
  labs(y = "Number of events", x = "Identified amino acid", tag = "d") + 
  theme_base(base_size = 22) + 
  theme(plot.background = element_blank(), 
        # axis.title.x = element_blank(), 
        plot.tag = element_text(face = "bold"), 
        legend.position = "none")
```


```{r}
Freq_N <- EAFNL[, .N, .(seq, Pred1, file_id)]
```

```{r}
Freq_N <- merge(Freq_N, IntervalTime, by.x = "Pred1", by.y = "A")
```

```{r fig.width=16, fig.height=8}
ggplot(Freq_N[seq == "EAFNL"], aes(x = Pred1, y = N/Median, fill = seq)) + 
  geom_col(position = position_dodge()) + 
  geom_text(aes(x = Pred1, y = N/Median * 1.3, label = round(N/Median)), position = position_dodge(width = 0.9), size = 6) + 
  scale_y_log10() + 
  scale_fill_wsj() + 
  facet_wrap(~ file_id, nrow = 2) + 
  labs(y = "Number of events", x = "Identified amino acid", title = "N / Standard AA median", tag = "EAFNL") + 
  theme_base(base_size = 22) + 
  theme(plot.background = element_blank(), 
        # axis.title.x = element_blank(), 
        plot.tag = element_text(face = "bold"), 
        legend.position = "none")
```

```{r fig.width=16, fig.height=8}
ggplot(Freq_N[seq == "LNFAE"], aes(x = Pred1, y = N/Median, fill = seq)) + 
  geom_col(position = position_dodge()) + 
  geom_text(aes(x = Pred1, y = N/Median * 1.3, label = round(N/Median)), position = position_dodge(width = 0.9), size = 6) + 
  scale_y_log10() + 
  scale_fill_wsj() + 
  facet_wrap(~ file_id, nrow = 2) + 
  labs(y = "Number of events", x = "Identified amino acid", title = "N / Standard AA median", tag = "LNFAE") + 
  theme_base(base_size = 22) + 
  theme(plot.background = element_blank(), 
        # axis.title.x = element_blank(), 
        plot.tag = element_text(face = "bold"), 
        legend.position = "none")
```
